fix:  1. 修复 windows端  右键发送 是直接发送了，没有出现使用md格式发送  2.其他bug修复

+++ app/Services/WebSocketService.php
@@ -172,8 +172,11 @@ class WebSocketService implements WebSocketHandlerInterface
      */
     public function onClose(Server $server, $fd, $reactorId)
     {
-        Task::deliver(new LineTask($this->getUserid($fd), false));  // 通知离线
-        $this->deleteUser($fd);
+        $userid = $this->getUserid($fd);
+        if($userid){
+            Task::deliver(new LineTask($userid, false));  // 通知离线
+            $this->deleteUser($fd);
+        }
     }
 
     /** ****************************************************************************** */
